+++ HISTORY.rst
@@ -6,6 +6,8 @@ History
 ------------------
 
 * Move all alpha related operations to predict
+* Assume default LinearRegression if estimator is None
+* Improve documentation
 
 0.1.3 (2021-04-30)
 ------------------